feat(bcgov/fieldset): update fieldset
update story to use mdx format

#294
