+++ packages/bcgov-theme/stories/fieldset/Fieldset.stories.mdx
@@ -1,0 +1,27 @@
+import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
+import Fieldset from '../../src/Fieldset';
+import * as stories from './Fieldset.stories.tsx';
+import { ArgsTable } from '@storybook/addon-docs/blocks';
+import { argTypes } from './args';
+
+<Meta title="Components/Fieldset" component={Fieldset} argTypes={argTypes} />
+
+# Fieldset
+
+The fieldset is used to group related items within a form.
+
+export const Template = (args) => <Fieldset {...args} />
+
+<Canvas>
+  <Story story={stories.Default} name="Default" />
+</Canvas>
+
+## Arguments
+
+<ArgsTable story="Default" />
+
+## HTML only
+
+<Canvas>
+  <Story story={stories.HTML} name="html" />
+</Canvas>
